# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Авито Питомец — API",
    "version": "1.0.0",
    "description": "Контракт фронт ↔ бэк.\n\nПРАВИЛА, ОБЯЗАТЕЛЬНЫЕ ДЛЯ ОБЕИХ СТОРОН:\n1. Клиент НЕ передаёт XP, уровень, прогресс задач, значения показателей. Только факт действия и его идентификатор.\n2. Все расчёты (decay, XP, уровень, настроение, стрик, право на награду) — на сервере.\n3. Порядок множителей: XP = базовый × множитель_настроения × множитель_стрика. Настроение берётся ПОСЛЕ изменения показателя. Округление вниз до целого.\n4. Показатели целые 0..100, потолок жёсткий: 82 + 30 = 100, остаток сгорает.\n5. Суточный потолок ухода обрезает последнее действие ЧАСТИЧНО (осталось 5 из 40 — начисляем 5, не 0).\n6. Сутки и стрик считаются по timezone аккаунта, а не по времени устройства.\n7. Все ответы в конверте { data, meta }.",
    "contact": { "name": "Фронтенд", "email": "frontend@avito-pet.local" }
  },
  "servers": [
    { "url": "https://api.avito-pet.local/api/v1", "description": "прод" },
    { "url": "http://localhost:8080/api/v1", "description": "локально через docker-compose" }
  ],
  "security": [ { "bearerAuth": [] } ],
...
  "paths": {
...
    "/summary/daily": {
      "get": {
        "tags": ["social"],
        "summary": "Сводка дня",
        "description": "Решение о показе принимает СЕРВЕР через поле shouldShow. Клиент не вычисляет «наступил ли вечер»: перевод часов на устройстве сломал бы логику.",
        "parameters": [ { "name": "date", "in": "query", "required": false,
          "schema": { "type": "string", "format": "date" } } ],
        "responses": { "200": { "description": "Сводка", "content": { "application/json": { "schema": {
          "allOf": [ { "$ref": "#/components/schemas/Envelope" },
            { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/DailySummary" } } } ] } } } } }
      }
    },
    "/summary/daily/seen": {
      "post": {
        "tags": ["social"], "summary": "Отметить сводку просмотренной",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["date"],
          "properties": { "date": { "type": "string", "format": "date" } } } } } },
        "responses": { "204": { "description": "Записано" } }
      }
    },
    "/leaderboard": {
      "get": {
        "tags": ["social"],
        "summary": "Лидерборд",
        "description": "Поле me приходит ОТДЕЛЬНО от items, независимо от страницы: иначе свою строку пришлось бы искать по всем страницам. Ранжирование по XP за 7 дней, тай-брейк — длина стрика. Отдаются только публичные поля.",
        "parameters": [
          { "name": "scope", "in": "query", "required": true,
            "schema": { "type": "string", "enum": ["league", "friends", "top"] } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 25, "maximum": 50 } }
        ],
        "responses": { "200": { "description": "Страница лидерборда", "content": { "application/json": { "schema": {
          "allOf": [ { "$ref": "#/components/schemas/Envelope" },
            { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/LeaderboardPage" } } } ] } } } } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
    "schemas": {
      "Envelope": {
        "type": "object", "required": ["meta"],
        "properties": {
          "data": { "description": "Полезная нагрузка. null при ошибке" },
          "meta": { "$ref": "#/components/schemas/Meta" }
        }
      },
      "Meta": {
        "type": "object", "required": ["code", "message"],
        "properties": {
          "code": { "type": "integer", "description": "ВСЕГДА число, дублирует HTTP-статус", "examples": [200] },
          "message": { "type": "string", "description": "Текст для пользователя, на русском" },
          "error": { "$ref": "#/components/schemas/ErrorCode" },
          "requestId": { "type": "string", "format": "uuid",
            "description": "Эхо заголовка X-Request-Id. По нему бэк находит лог по скриншоту бага" },
          "retryAfterSec": { "type": "integer", "description": "Только при 429" }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "Машинный код. Фронт ветвится ТОЛЬКО по нему: message перепишут при вычитке текстов, и клиент молча сломается.",
        "enum": [
          "VALIDATION_ERROR", "UNAUTHORIZED", "TOKEN_EXPIRED", "NOT_FOUND",
          "PET_ALREADY_EXISTS", "PET_IS_SLEEPING", "ACTION_LIMIT_REACHED", "ACTION_RATE_LIMITED",
          "EVENT_REJECTED", "REWARD_NOT_ELIGIBLE", "REWARD_ALREADY_CLAIMED", "REWARD_EXPIRED",
          "REWARD_NOT_CLAIMED", "COSMETIC_NOT_OWNED", "INTERNAL_ERROR"
        ]
      },
...
      "PresetId": { "type": "string", "enum": ["green", "blue", "purple"] },
...
      "Stats": {
        "type": "object", "required": ["hunger", "joy", "clean", "energy"],
        "description": "Целые 0..100. Потолок жёсткий: 82 + 30 = 100, остаток сгорает",
        "properties": {
          "hunger": { "type": "integer", "minimum": 0, "maximum": 100 },
          "joy":    { "type": "integer", "minimum": 0, "maximum": 100 },
          "clean":  { "type": "integer", "minimum": 0, "maximum": 100 },
          "energy": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
...
      "PetMood": {
        "type": "string",
        "description": "Производная от МИНИМАЛЬНОГО показателя. Клиент её не вычисляет",
        "enum": ["radiant", "happy", "neutral", "sad", "sick", "sleeping"]
      },
...
      "TaskClass": {
        "type": "string",
        "description": "care — внутри игры; setup — нулевой след у других людей (сохранить поиск, обновить фото своего объявления); explore — слабый след (просмотры, сравнение, избранное); weekly — недельная",
        "enum": ["care", "setup", "explore", "weekly"]
      },
      "Task": {
        "type": "object",
        "required": ["id", "class", "title", "icon", "xpReward", "progress", "status", "expiresAt"],
        "properties": {
          "id": { "type": "string" },
          "class": { "$ref": "#/components/schemas/TaskClass" },
          "title": { "type": "string", "examples": ["Посмотри 3 объявления"] },
          "subtitle": { "type": "string", "examples": ["в разделе «Электроника»"] },
          "icon": { "type": "string", "description": "Ключ иконки, фронт рисует SVG сам" },
          "xpReward": { "type": "integer" },
          "progress": { "type": "object", "required": ["current", "target"],
            "properties": { "current": { "type": "integer" }, "target": { "type": "integer" } } },
          "status": { "type": "string", "enum": ["active", "completed", "expired"] },
          "deepLink": { "type": "string", "format": "uri" },
          "requiresEvent": { "type": "array", "items": { "$ref": "#/components/schemas/ActivityType" } },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ActivityType": {
        "type": "string",
        "description": "Первые девять могут быть задачами. Последние четыре — только признания: дают XP, но никогда не выдаются как задача, иначе продавцы получат спам",
        "enum": [
          "listing.viewed", "listing.compared", "listing.favorited",
          "search.saved", "profile.updated", "ad.photo_updated", "ad.delivery_enabled",
          "autoteka.report_viewed", "delivery.section_opened",
          "seller.contacted", "ad.published", "deal.completed", "review.left"
        ]
      },
...
      "Streak": {
        "type": "object", "required": ["current", "best", "freezes", "todayCounted", "multiplier"],
        "properties": {
          "current": { "type": "integer" },
          "best": { "type": "integer", "description": "Никогда не уменьшается" },
          "freezes": { "type": "integer", "minimum": 0, "maximum": 2 },
          "todayCounted": { "type": "boolean" },
          "multiplier": { "type": "number", "examples": [1.2],
            "description": "1–6 дн ×1.0 · 7–13 ×1.1 · 14–29 ×1.2 · 30–99 ×1.3 · 100–499 ×1.4 · 500+ ×1.5" },
          "frozenUsed": { "type": "boolean", "description": "Заморозка списана автоматически — сказать честно в сводке" },
          "broken": { "type": "boolean" },
          "nextMilestone": { "type": "object",
            "properties": { "days": { "type": "integer" }, "title": { "type": "string" },
                            "daysLeft": { "type": "integer" }, "rewardId": { "type": "string" } } }
        }
      },
...
      "RewardTier": {
        "type": "string",
        "description": "cosmetic — себестоимость 0, выдаётся напрямую · soft — бонусы кошелька, мелкие скидки · premium — платные услуги Авито, только с 10 уровня и вехи 100-дневного стрика. premium сегодня не встречается ни в одной награде каталога: вехи 100 без стрик-системы не существует, а выдумывать недостижимую награду нечестнее, чем не показывать её вовсе (docs/DECISIONS.md → 10.08)",
        "enum": ["cosmetic", "soft", "premium"]
      },
      "RewardStatus": { "type": "string", "enum": ["locked", "available", "claimed", "used", "expired"] },
      "Reward": {
        "type": "object",
        "required": ["id", "tier", "title", "conditionLabel", "progress", "status"],
        "description": "Энтайтлмент, не промокод (docs/DECISIONS.md → 10.08): право на бонус привязано к user_id в reward_grants, кода, который можно переслать, в ответе нет вообще.",
        "properties": {
          "id": { "type": "string" },
          "tier": { "$ref": "#/components/schemas/RewardTier" },
          "title": { "type": "string", "examples": ["Скидка на Авито Доставку"] },
          "description": { "type": "string" },
          "conditionLabel": { "type": "string", "examples": ["Уровень 10"],
            "description": "ГОТОВЫЙ текст с сервера. Фронт не склеивает фразы из чисел" },
          "progress": { "type": "object", "required": ["current", "target", "unit"],
            "properties": { "current": { "type": "integer" }, "target": { "type": "integer" },
              "unit": { "type": "string", "enum": ["level", "day", "task"] } } },
          "etaDays": { "type": "integer", "description": "Оценка сервера по темпу пользователя. Отсутствует, если сигнала недостаточно для честной оценки — не выдумывается" },
          "status": { "$ref": "#/components/schemas/RewardStatus" },
          "cosmeticId": { "type": "string", "description": "Если tier=cosmetic" },
          "claimedAt": { "type": "string", "format": "date-time" },
          "usedAt": { "type": "string", "format": "date-time",
            "description": "Когда награда применена (POST /rewards/{rewardId}/redeem-click). Отсутствует, пока status не used" },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
      "DailySummary": {
        "type": "object",
        "properties": {
          "date": { "type": "string", "format": "date" },
          "shouldShow": { "type": "boolean", "description": "Решает сервер, не клиент" },
          "xpTotal": { "type": "integer" },
          "multiplier": { "type": "number" },
          "breakdown": { "type": "array", "items": { "type": "object",
            "properties": { "key": { "type": "string" }, "label": { "type": "string" },
                            "count": { "type": "integer" }, "xp": { "type": "integer" } } } },
          "streak": { "$ref": "#/components/schemas/Streak" },
          "pet": { "type": "object",
            "properties": {
              "levelBefore": { "type": "integer" }, "levelAfter": { "type": "integer" },
              "moodBefore": { "$ref": "#/components/schemas/PetMood" },
              "moodAfter": { "$ref": "#/components/schemas/PetMood" },
              "statsAfter": { "$ref": "#/components/schemas/Stats" } } },
          "tomorrow": { "type": "object",
            "properties": {
              "tasks": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } },
              "streakWillBe": { "type": "integer" },
              "multiplierWillBe": { "type": "number" },
              "nextReward": { "$ref": "#/components/schemas/Reward" } } },
          "aiNote": { "type": "string",
            "description": "Персональная фраза от питомца. Генерируется на бэке: ключи LLM не должны попасть в браузер" }
        }
      },
      "LeaderboardEntry": {
        "type": "object",
        "properties": {
          "rank": { "type": "integer" },
          "userId": { "type": "string" },
          "nickname": { "type": "string" },
          "presetId": { "$ref": "#/components/schemas/PresetId" },
          "level": { "type": "integer" },
          "weeklyXp": { "type": "integer" },
          "streak": { "type": "integer" },
          "isMe": { "type": "boolean" }
        }
      },
      "LeaderboardPage": {
        "type": "object",
        "properties": {
          "scope": { "type": "string", "enum": ["league", "friends", "top"] },
          "league": { "type": "object",
            "properties": { "id": { "type": "string" }, "name": { "type": "string" },
              "endsAt": { "type": "string", "format": "date-time" },
              "promoteCount": { "type": "integer" }, "relegateCount": { "type": "integer" } } },
          "me": { "$ref": "#/components/schemas/LeaderboardEntry" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/LeaderboardEntry" } },
          "nextCursor": { "type": "string" }
        }
      },
...
    }
  },
...
}